// src/api/types.gen.ts
export type AddPurchaseEntry = {
    name: string;
    cost: number;
    participantId: string;
    purchaseUsages: Array<UpsertPurchaseUsageEntry>;
};
// ...
export type GetPurchaseEntry = {
    id: string;
    name: string;
    cost: number;
    isComplete: boolean;
    eventId: string;
    purchaseUsages?: Array<PurchaseUsageShortEntry> | null;
    participant?: ParticipantShortEntry;
};
// ...
export type ParticipantShortEntry = {
    id: string;
    name?: string | null;
    imageUri?: string | null;
};

export type ParticipantShortEntryResponse = {
    data?: Array<ParticipantShortEntry> | null;
    totalPages?: number;
    total?: number;
};

export type PurchaseShortEntry = {
    id: string;
    eventId: string;
    name: string;
    cost: number;
    isComplete: boolean;
};

export type PurchaseUsageShortEntry = {
    purchaseUsageId: string;
    purchaseId?: string;
    participantId?: string;
    amount?: number;
};
// ...
export type UpdatePurchaseEntry = {
    name: string;
    cost: number;
    purchaseUsages: Array<UpsertPurchaseUsageEntry>;
};

export type UpsertParticipantEntry = {
    userId: string;
};

export type UpsertPurchaseUsageEntry = {
    participantId: string;
    amount: number;
};